batbot/batbot_bringup/requirements.txt
# Author: Ben Westcott
# Created: 4/22/23

matplotlib
numpy
pyserial
pyyaml
pygnssutils
gpxpy
scipy
pynmeagps
pyrtcm
pyubx2
paho-mqtt
PyQt6
qdarkstyle
pyqtgraph
pybind11
setuptools

# for building docs
sphinx
sphinx-rtd-theme
sphinx-copybutton
sphinx-new-tab-link